* origin/develop-1.10: (27 commits)
  Backport changes to standard actions.  Use CorbaUtil so these operations are interruptable.
  refs #1782 Use new CorbaUtil method to invoke operations in interruptable background manor.
  refs #1778 Added BulkIO Preference page under REDHAWK category.  Use the preference value if no system property is defined.  Allow the current system property to override the preference value.
  refs #1777 Implementing FrontendAction subclasses for Properties view toolbar buttons so they will work in e4
  Invoke deallocate in CorbaUtil call.
  refs #1775 null check added to TunerUtils.getControlId()
  Use fields in super class instead of overriding methods.
  Make release action use new CorbaUtils method.
  Use new CorbaUtil method in deallocate call.
  Fixed context menu
  Refactored the Property Viewer to be reusable by the frontend section.  FrontEnd Property section now uses same controls as the property control.
  refs #1768 Fixing typo in previous commit
  refs #1768 Fixing FrontEnd properties section to show correct toolbar buttons when unallocated tuner container is selected
  Removing commented-out lines that are no longer needed
  Mark SCA Event Channel name as a read only property field.
  Ensure we call set value with the new value from the notification.  There is a chance the tuner status value will be set back to the old value by the refreshing before the job runs.
  Updated IDL libraries.
  fixed creeping column issue for FrontEnd properites sections
  Edits to properly handle device updates from the FrontEnd section
  Null safe check on allocate handler.
  ...

+++ plugins/gov.redhawk.frontend/src-model/gov/redhawk/frontend/TunerStatus.java
@@ -718,12 +718,12 @@ public interface TunerStatus extends EObject {
 	 * @return the value of the '<em>Reference Source</em>' attribute.
 	 * @see #isSetReferenceSource()
 	 * @see #unsetReferenceSource()
-	 * @see #setReferenceSource(long)
+	 * @see #setReferenceSource(int)
 	 * @see gov.redhawk.frontend.FrontendPackage#getTunerStatus_ReferenceSource()
 	 * @model unique="false" unsettable="true"
 	 * @generated
 	 */
-	long getReferenceSource();
+	int getReferenceSource();
 
 	/**
 	 * Sets the value of the '{@link gov.redhawk.frontend.TunerStatus#getReferenceSource <em>Reference Source</em>}'
@@ -736,7 +736,7 @@ public interface TunerStatus extends EObject {
 	 * @see #getReferenceSource()
 	 * @generated
 	 */
-	void setReferenceSource(long value);
+	void setReferenceSource(int value);
 
 	/**
 	 * Unsets the value of the '{@link gov.redhawk.frontend.TunerStatus#getReferenceSource <em>Reference Source</em>}'
@@ -745,7 +745,7 @@ public interface TunerStatus extends EObject {
 	 * <!-- end-user-doc -->
 	 * @see #isSetReferenceSource()
 	 * @see #getReferenceSource()
-	 * @see #setReferenceSource(long)
+	 * @see #setReferenceSource(int)
 	 * @generated
 	 */
 	void unsetReferenceSource();
@@ -758,7 +758,7 @@ public interface TunerStatus extends EObject {
 	 * @return whether the value of the '<em>Reference Source</em>' attribute is set.
 	 * @see #unsetReferenceSource()
 	 * @see #getReferenceSource()
-	 * @see #setReferenceSource(long)
+	 * @see #setReferenceSource(int)
 	 * @generated
 	 */
 	boolean isSetReferenceSource();
